Separate charts in PlayerScoreChart
Fixes #34

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/josve/mahjong/issues/34?shareId=XXXX-XXXX-XXXX-XXXX).
